YUNKE-3283 【Server】作为云视课堂运营者，我希望所有学校能在一套系统中运行，而不是每个学校一套部署

+++ value_int64.go
@@ -4,7 +4,7 @@ var _ Value = (*Int64Value)(nil)
 
 // Int64Value 整形值
 type Int64Value struct {
-	value int
+	value int64
 }
 
 func (iv *Int64Value) Value() interface{} {
